+++ packages/cheqd-blockchain-api/src/index.js
@@ -228,7 +228,7 @@ export class CheqdAPI extends AbstractApiProvider {
     const sender = from ?? (await this.sdk.options.wallet.getAccounts())[0].address;
     const payment = {
       amount: [amount],
-      gas: '1200000', // TODO: dynamically calculate needed amount
+      gas: '1400000', // TODO: dynamically calculate needed amount
       payer: sender,
     };