+++ packages/cheqd-blockchain-modules/src/common/inject-public-keys.js
@@ -3,17 +3,14 @@ import {
   u8aToString,
   withExtendedStaticProperties,
 } from '@docknetwork/credential-sdk/utils';
-import {
-  CheqdCreateResource,
-  CheqdDid,
-} from '@docknetwork/credential-sdk/types';
+import { CheqdCreateResource } from '@docknetwork/credential-sdk/types';
 import createInternalCheqdModule from './create-internal-cheqd-module';
 import { validateResource } from './resource';
 
 const methods = {
   addPublicKey(id, publicKey, did) {
     return new CheqdCreateResource(
-      CheqdDid.from(did).value.value,
+      this.types.Did.from(did).value.value,
       TypedUUID.from(id),
       '1.0',
       [],