+++ packages/dock-blockchain-modules/tests/integration/modules/blob-module.test.js
@@ -1,7 +1,10 @@
 import { DockAPI } from "@docknetwork/dock-blockchain-api";
 import { DockDid } from "@docknetwork/credential-sdk/types";
 import generateBlobModuleTests from "@docknetwork/credential-sdk/modules/tests/blob-module";
-import { MultiApiBlobModule, MultiApiDIDModule } from "@docknetwork/credential-sdk/modules";
+import {
+  MultiApiBlobModule,
+  MultiApiDIDModule,
+} from "@docknetwork/credential-sdk/modules";
 import { DockDIDModule, DockBlobModule } from "../../../src";
 import {
   FullNodeEndpoint,
@@ -28,11 +31,14 @@ describe("BlobModule", () => {
 
   generateBlobModuleTests(
     { did: new DockDIDModule(dock), blob: new DockBlobModule(dock) },
-    { DID: DockDid, BlobId: DockBlobId }
+    { Did: DockDid, BlobId: DockBlobId }
   );
 
   generateBlobModuleTests(
-    { did: new MultiApiDIDModule([new DockDIDModule(dock)]), blob: new MultiApiBlobModule([new DockBlobModule(dock)]) },
-    { DID: DockDid, BlobId: DockBlobId }
+    {
+      did: new MultiApiDIDModule([new DockDIDModule(dock)]),
+      blob: new MultiApiBlobModule([new DockBlobModule(dock)]),
+    },
+    { Did: DockDid, BlobId: DockBlobId }
   );
 });